+++ src/store/obstacles/index.ts
@@ -2,5 +2,5 @@ export { default as Tree } from "./tree";
 export { default as Bush } from "./bush";
 export { default as Crate } from "./crate";
 export { default as Stone } from "./stone";
-export { default as mosin_Tree } from "./mosin_tree";
-export { default as soviet_Crate} from "./soviet_crate"
+export { default as MosinTree } from "./mosin_tree";
+export { default as SovietCrate } from "./soviet_crate"
