+++ src/store/weapons/guns/m1100.ts
@@ -13,6 +13,5 @@ export default class M1100 extends GunWeapon {
 	ticks = 50;
 	delay = 20;
 	recoil = 0.4;
-	spread = 9;
-	bullets = 4;
+	bullets = 9;
 }
