+++ src/store/weapons/guns/m9.ts
@@ -13,6 +13,4 @@ export default class M9 extends GunWeapon {
 	ticks = 50;
 	delay = 12;
 	recoil = 0.1;
-	spread = 1;
-	bullets = 15;
 }
