+++ src/store/weapons/guns/mp220.ts
@@ -1,18 +1,17 @@
 import { GunWeapon } from "../../../types/weapon";
 import { GunColor } from "../../../types/misc";
 
-export default class M870 extends GunWeapon {
-	id = "M870";
-	name = "M870";
-	continuous = false;
+export default class MP220 extends GunWeapon {
+	id = "MP220";
+	name = "MP220";
+	continuous = true;
 	ammo = GunColor.RED;
 	speed = 0.8;
 	accuracy = 1.0;
 	inaccuracy = 0.05;
 	weight = 1.0;
 	ticks = 50;
-	delay = 50;
-	recoil = 1.0;
-	spread = 6;
-	bullets = 5;
+	delay = 10;
+	recoil = 0.01;
+	bullets = 8;
 }
