+++ src/store/weapons/index.ts
@@ -1,7 +1,7 @@
 export { default as Fists } from "./fists";
 export { default as M9 } from "./guns/m9";
-export { default as mosin_nagant} from "./guns/mosin_nagant"
-export { default as M870} from "./guns/M870";
-export { default as MP220} from "./guns/MP220";
-export { default as M1100} from "./guns/M1100";
-export { default as mp5} from "./guns/mp5"
+export { default as MosinNagant } from "./guns/mosin_nagant"
+export { default as M870 } from "./guns/m870";
+export { default as MP220 } from "./guns/mp220";
+export { default as M1100 } from "./guns/m1100";
+export { default as mp5 } from "./guns/mp5"
